backend/requirements.txt
Flask
Flask-Cors
supabase
python-dotenv
playwright
requests
PyGithub
groq==0.5.0
httpx<0.28
google-generativeai>=0.7.0
gunicorn==21.2.0